RHCLOUD-35642: Port setting additional environment variables for IQE (#114)
This is needed to allow running in Konflux with additional environment variables.

Logic is ported away from using bash to something more portable with awk. In addition this should be easier to maintain because the purpose is more clear and the logic is easier to follow.

Comment included from original script so others know this is a comma delimited string.